feat(fx plugins): put plugin.Serve logic into dedicated function
