+++ internal/connectors/plugins/grpc_server.go
@@ -2,20 +2,17 @@ package plugins
 
 import (
 	"context"
-	"fmt"
 	"sync"
 
-	"github.com/formancehq/go-libs/otlp/otlpmetrics"
-	"github.com/formancehq/go-libs/otlp/otlptraces"
-	"github.com/formancehq/go-libs/service"
 	"github.com/formancehq/payments/internal/connectors/grpc"
 	"github.com/formancehq/payments/internal/models"
 	"github.com/hashicorp/go-plugin"
-	"github.com/spf13/cobra"
 	"go.uber.org/fx"
 )
 
-type Server interface{}
+type Server interface {
+	Serve(wg *sync.WaitGroup, shutdowner fx.Shutdowner)
+}
 
 type server struct {
 	plugin models.Plugin
@@ -27,21 +24,7 @@ func NewServer(lc fx.Lifecycle, shutdowner fx.Shutdowner, plg models.Plugin) Ser
 	lc.Append(fx.Hook{
 		OnStart: func(ctx context.Context) error {
 			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				plugin.Serve(&plugin.ServeConfig{
-					HandshakeConfig: grpc.Handshake,
-					Plugins: map[string]plugin.Plugin{
-						"psp": &grpc.PSPGRPCPlugin{Impl: NewGRPCImplem(srv.plugin)},
-					},
-
-					// A non-nil value here enables gRPC serving for this plugin...
-					GRPCServer: plugin.DefaultGRPCServer,
-				})
-				// when Serve has ended the server closed usually because the plugin.Client told it to
-				// if the parent application (managed by fx) is still running we need to tell it the plugin is done
-				shutdowner.Shutdown(fx.ExitCode(0))
-			}()
+			go srv.Serve(wg, shutdowner)
 			return nil
 		},
 		OnStop: func(ctx context.Context) error {
@@ -54,29 +37,21 @@ func NewServer(lc fx.Lifecycle, shutdowner fx.Shutdowner, plg models.Plugin) Ser
 	return srv
 }
 
-// TODO(polo): metrics
-func NewPlugin(name string, pluginFn models.PluginFn) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:          fmt.Sprintf("serve %s plugin", name),
-		Aliases:      []string{name},
-		Short:        fmt.Sprintf("Launch %s plugin server", name),
-		SilenceUsage: true,
-		RunE:         runServer(pluginFn),
-	}
+func (s *server) Serve(wg *sync.WaitGroup, shutdowner fx.Shutdowner) {
+	defer func() {
+		wg.Done()
+		// when Serve has ended the server closed usually because the plugin.Client told it to
+		// if the parent application (managed by fx) is still running we need to tell it the plugin is done
+		shutdowner.Shutdown(fx.ExitCode(0))
+	}()
 
-	service.AddFlags(cmd.Flags())
-	otlpmetrics.AddFlags(cmd.Flags())
-	otlptraces.AddFlags(cmd.Flags())
-	return cmd
-}
+	plugin.Serve(&plugin.ServeConfig{
+		HandshakeConfig: grpc.Handshake,
+		Plugins: map[string]plugin.Plugin{
+			"psp": &grpc.PSPGRPCPlugin{Impl: NewGRPCImplem(s.plugin)},
+		},
 
-func runServer(pluginFn models.PluginFn) func(cmd *cobra.Command, args []string) error {
-	return func(cmd *cobra.Command, args []string) error {
-		// TODO initialise logger here?
-		opts := fx.Options(
-			fx.Provide(pluginFn, NewServer),
-			fx.Invoke(func(Server) {}),
-		)
-		return service.New(cmd.OutOrStderr(), opts).Run(cmd)
-	}
+		// A non-nil value here enables gRPC serving for this plugin...
+		GRPCServer: plugin.DefaultGRPCServer,
+	})
 }
